+++ packages/playwright-tests/cli-optimization.spec.js
@@ -71,12 +71,14 @@ for (let { port, name } of test_variants) {
   // delivered as a classic script with its body intact.
   test(`classic js asset with_minify(false) is copied byte-for-byte in ${name}`, async ({ page }) => {
     await page.goto(`http://localhost:${port}`);
+    // @ts-ignore
     const value = await page.evaluate(() => window.__iife_classic_value);
     expect(value).toBe("ok-classic");
   });
 
   test(`classic js asset with_minify(true) stays a classic script in ${name}`, async ({ page }) => {
     await page.goto(`http://localhost:${port}`);
+    // @ts-ignore
     const value = await page.evaluate(() => window.__iife_minify_value);
     expect(value).toBe("ok-minify");
   });
@@ -86,6 +88,7 @@ for (let { port, name } of test_variants) {
   // which only parses when the script tag has `type="module"`.
   test(`js asset with_module(true) is loaded as an ES module in ${name}`, async ({ page }) => {
     await page.goto(`http://localhost:${port}`);
+    // @ts-ignore
     const value = await page.evaluate(() => window.__esm_module_value);
     expect(value).toBe("ok-module");
     const moduleScript = page.locator('script[type="module"][src*="esm_module"]');
@@ -96,10 +99,12 @@ for (let { port, name } of test_variants) {
   // `export` and emit `<script type="module">` even without `with_module(true)`.
   test(`js asset with top-level export is auto-detected as ES module in ${name}`, async ({ page }) => {
     await page.goto(`http://localhost:${port}`);
+    // @ts-ignore
     const value = await page.evaluate(() => window.__esm_auto_value);
     expect(value).toBe("ok-auto");
     // import.meta only parses inside a real module; if the script were emitted
     // as classic the file would have errored at parse time.
+    // @ts-ignore
     const metaType = await page.evaluate(() => window.__esm_auto_meta);
     expect(metaType).toBe("string");
     const moduleScript = page.locator('script[type="module"][src*="esm_auto"]');
